move immutable style to api-types module
