+++ core/common/api/src/main/java/org/example/age/common/api/data/internal/ApiObjectMapperModule.java
@@ -5,7 +5,7 @@ import dagger.Module;
 import dagger.Provides;
 import javax.inject.Named;
 import javax.inject.Singleton;
-import org.example.age.data.utils.DataMapper;
+import org.example.age.data.mapper.DataMapper;
 
 /** Dagger module that publishes a binding for <code>@Named("api") {@link ObjectMapper}</code>. */
 @Module
