+++ core/common/service-types/build.gradle.kts
@@ -6,6 +6,7 @@ plugins {
 dependencies {
     annotationProcessor("org.immutables:value")
 
+    api(project(":api-types"))
     api(project(":core:data"))
     api("com.fasterxml.jackson.core:jackson-databind")
     api("org.immutables:value-annotations")
