+++ core/data/src/main/java/org/example/age/data/certificate/VerificationSession.java
@@ -1,15 +1,13 @@
 package org.example.age.data.certificate;
 
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
-import com.fasterxml.jackson.databind.annotation.JsonSerialize;
+import org.example.age.api.ApiStyle;
 import org.example.age.data.crypto.Aes256Key;
-import org.example.age.data.utils.DataStyle;
 import org.immutables.value.Value;
 
 /** Session to pseudonymously verify a person's age (and guardians, if applicable). */
 @Value.Immutable
-@DataStyle
-@JsonSerialize(as = ImmutableVerificationSession.class)
+@ApiStyle
 @JsonDeserialize(as = ImmutableVerificationSession.class)
 public interface VerificationSession {
