+++ core/data/src/main/java/org/example/age/data/crypto/AesGcmEncryptionPackage.java
@@ -1,16 +1,14 @@
 package org.example.age.data.crypto;
 
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
-import com.fasterxml.jackson.databind.annotation.JsonSerialize;
 import java.util.Optional;
+import org.example.age.api.ApiStyle;
 import org.example.age.data.crypto.internal.EncryptionUtils;
-import org.example.age.data.utils.DataStyle;
 import org.immutables.value.Value;
 
 /** Encryption package for plaintext encrypted using AES-256/GCM. */
 @Value.Immutable
-@DataStyle
-@JsonSerialize(as = ImmutableAesGcmEncryptionPackage.class)
+@ApiStyle
 @JsonDeserialize(as = ImmutableAesGcmEncryptionPackage.class)
 public interface AesGcmEncryptionPackage {
