+++ core/data/src/main/java/org/example/age/data/mapper/DataMapper.java
@@ -1,4 +1,4 @@
-package org.example.age.data.utils;
+package org.example.age.data.mapper;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.datatype.guava.GuavaModule;
@@ -6,7 +6,7 @@ import com.fasterxml.jackson.datatype.guava.GuavaModule;
 /** {@link ObjectMapper} for all types that can be serialized as JSON. */
 public final class DataMapper {
 
-    private static final ObjectMapper mapper = createObjectMapper();
+    private static final ObjectMapper mapper = create();
 
     /** Gets the {@link ObjectMapper} singleton. */
     public static ObjectMapper get() {
@@ -14,7 +14,7 @@ public final class DataMapper {
     }
 
     /** Creates the {@link ObjectMapper} singleton. */
-    private static ObjectMapper createObjectMapper() {
+    private static ObjectMapper create() {
         ObjectMapper mapper = new ObjectMapper();
         mapper.registerModule(new GuavaModule());
         return mapper;
