+++ core/data/src/main/java/org/example/age/data/user/VerifiedUser.java
@@ -1,10 +1,9 @@
 package org.example.age.data.user;
 
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
-import com.fasterxml.jackson.databind.annotation.JsonSerialize;
 import java.util.List;
+import org.example.age.api.ApiStyle;
 import org.example.age.data.crypto.SecureId;
-import org.example.age.data.utils.DataStyle;
 import org.immutables.value.Value;
 
 /**
@@ -13,8 +12,7 @@ import org.immutables.value.Value;
  * <p>The only PII a {@link VerifiedUser} contains is an age, which can be anonymized into an age range.</p>
  */
 @Value.Immutable
-@DataStyle
-@JsonSerialize(as = ImmutableVerifiedUser.class)
+@ApiStyle
 @JsonDeserialize(as = ImmutableVerifiedUser.class)
 public interface VerifiedUser {
