+++ core/data/src/test/java/org/example/age/data/certificate/AgeCertificateTest.java
@@ -10,8 +10,8 @@ import java.util.Optional;
 import org.example.age.data.crypto.Aes256Key;
 import org.example.age.data.crypto.AesGcmEncryptionPackage;
 import org.example.age.data.crypto.SecureId;
+import org.example.age.data.mapper.DataMapper;
 import org.example.age.data.user.VerifiedUser;
-import org.example.age.data.utils.DataMapper;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
