+++ core/data/src/test/java/org/example/age/data/certificate/VerificationSessionTest.java
@@ -5,7 +5,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import com.fasterxml.jackson.core.type.TypeReference;
 import java.io.IOException;
 import java.time.Duration;
-import org.example.age.data.utils.DataMapper;
+import org.example.age.data.mapper.DataMapper;
 import org.junit.jupiter.api.Test;
 
 public final class VerificationSessionTest {
