+++ core/data/src/test/java/org/example/age/data/user/VerifiedUserTest.java
@@ -6,7 +6,7 @@ import com.fasterxml.jackson.core.type.TypeReference;
 import java.io.IOException;
 import java.util.List;
 import org.example.age.data.crypto.SecureId;
-import org.example.age.data.utils.DataMapper;
+import org.example.age.data.mapper.DataMapper;
 import org.junit.jupiter.api.Test;
 
 public final class VerifiedUserTest {
