+++ core/legacy/src/main/java/org/example/age/common/avs/config/AvsConfig.java
@@ -2,12 +2,12 @@ package org.example.age.common.avs.config;
 
 import java.security.PrivateKey;
 import java.time.Duration;
-import org.example.age.data.utils.DataStyle;
+import org.example.age.api.ApiStyle;
 import org.immutables.value.Value;
 
 /** Configuration for the age verification service. */
 @Value.Immutable
-@DataStyle
+@ApiStyle
 public interface AvsConfig {
 
     /** Creates a builder for the AVS configuration. */
