+++ core/legacy/src/main/java/org/example/age/common/avs/config/RegisteredSiteConfig.java
@@ -1,17 +1,15 @@
 package org.example.age.common.avs.config;
 
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
-import com.fasterxml.jackson.databind.annotation.JsonSerialize;
+import org.example.age.api.ApiStyle;
 import org.example.age.common.service.config.SiteLocation;
 import org.example.age.data.crypto.SecureId;
 import org.example.age.data.user.AgeThresholds;
-import org.example.age.data.utils.DataStyle;
 import org.immutables.value.Value;
 
 /** Configuration for a registered site. */
 @Value.Immutable
-@DataStyle
-@JsonSerialize(as = ImmutableRegisteredSiteConfig.class)
+@ApiStyle
 @JsonDeserialize(as = ImmutableRegisteredSiteConfig.class)
 public interface RegisteredSiteConfig {
