+++ core/legacy/src/test/java/org/example/age/common/avs/verification/VerificationManagerTest.java
@@ -30,9 +30,9 @@ import org.example.age.common.service.config.SiteLocation;
 import org.example.age.common.service.store.InMemoryPendingStoreFactoryModule;
 import org.example.age.data.certificate.VerificationSession;
 import org.example.age.data.crypto.SecureId;
+import org.example.age.data.mapper.DataMapper;
 import org.example.age.data.user.AgeThresholds;
 import org.example.age.data.user.VerifiedUser;
-import org.example.age.data.utils.DataMapper;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
