+++ core/site/service/src/main/java/org/example/age/site/service/config/SiteConfig.java
@@ -1,13 +1,13 @@
 package org.example.age.site.service.config;
 
 import java.time.Duration;
+import org.example.age.api.ApiStyle;
 import org.example.age.common.service.config.AvsLocation;
-import org.example.age.data.utils.DataStyle;
 import org.immutables.value.Value;
 
 /** Configuration for a site. */
 @Value.Immutable
-@DataStyle
+@ApiStyle
 public interface SiteConfig {
 
     /** Creates a builder for the site configuration. */
